+++ wiki/ReleaseNotes.md
@@ -2,6 +2,23 @@
 All relevant changes are documented in this file. You can find more information about
 the versioning scheme [here](JoynrVersioning.md).
 
+# joynr 1.15.10
+
+## API-relevant Changes
+None.
+
+## Other Changes
+None.
+
+## Configuration Property Changes
+None.
+
+## Security Fixes
+None.
+
+## Bug Fixes
+* **[C++]** Fix message/memory corruption when UDS sending-queue-size is reached.
+
 # joynr 1.15.9
 
 ## API-relevant Changes